+++ src/main/java/com/Teletubbies/Apollo/auth/controller/UserController.java
@@ -3,7 +3,6 @@ package com.Teletubbies.Apollo.auth.controller;
 import com.Teletubbies.Apollo.auth.domain.ApolloUser;
 import com.Teletubbies.Apollo.auth.dto.MemberInfoResponse;
 import com.Teletubbies.Apollo.auth.service.UserService;
-import com.Teletubbies.Apollo.jwt.dto.UserLoginRequestDto;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 @Slf4j
@@ -19,15 +18,6 @@ public class UserController {
         Long savedUserId = userService.saveUser(memberInfoResponse);
         log.info("유저 저장 완료");
         return savedUserId.toString();
-    }
-    @PostMapping("/login/user")
-    public String login(@RequestBody UserLoginRequestDto userLoginRequestDto){
-        String userLogin = userLoginRequestDto.getUserLogin();
-        String userId = userLoginRequestDto.getUserId();
-        log.info("user login: " + userLogin);
-        log.info("user id: " + userId);
-        String login = userService.login(userLogin, userId);
-        return login;
     }
     @GetMapping("/login/find/{userId}")
     public String findUserById(@PathVariable Long userId){
