+++ src/main/java/com/Teletubbies/Apollo/auth/service/UserService.java
@@ -4,12 +4,8 @@ import com.Teletubbies.Apollo.auth.domain.ApolloUser;
 import com.Teletubbies.Apollo.auth.dto.MemberInfoResponse;
 import com.Teletubbies.Apollo.auth.repository.UserRepository;
 import com.Teletubbies.Apollo.core.exception.ApolloException;
-import com.Teletubbies.Apollo.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,8 +19,6 @@ import static com.Teletubbies.Apollo.core.exception.CustomErrorCode.NOT_FOUND_US
 @Transactional(readOnly = true)
 public class UserService {
     private final UserRepository userRepository;
-    private final JwtTokenProvider jwtTokenProvider;
-    private final AuthenticationManagerBuilder authenticationManagerBuilder;
     @Transactional
     public Long saveUser(MemberInfoResponse memberInfoResponse) {
         ApolloUser apolloUserToSave = memberInfoResponse.changeDTOtoObj(memberInfoResponse);
@@ -32,20 +26,6 @@ public class UserService {
             throw new ApolloException(DUPLICATED_USER_ERROR, "이미 존재하는 회원입니다");
         ApolloUser savedApolloUser = userRepository.save(memberInfoResponse.changeDTOtoObj(memberInfoResponse));
         return savedApolloUser.getId();
-    }
-    public String login(String userLogin, String userId){
-        // 1. github login(ID) / git user id(PW) 를 기반으로 Authentication 객체 생성
-        // 이때 authentication 는 인증 여부를 확인하는 authenticated 값이 false
-        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(userLogin, userId);
-        log.info("authenticationtoken name: " + authenticationToken.getName());
-        log.info("authenticationtoken principal: " +authenticationToken.getPrincipal().toString());
-        log.info("authenticationtoken credentials " +authenticationToken.getCredentials().toString());
-
-        // 2. 실제 검증 (사용자 비밀번호 체크)이 이루어지는 부분
-        // authenticate 매서드가 실행될 때 CustomUserDetailsService 에서 만든 loadUserByUsername 메서드가 실행
-        //Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
-        //log.info("실제 검증 완료");
-        return "ok";
     }
     public ApolloUser getUserById(Long id){
         Optional<ApolloUser> optionalUser = userRepository.findById(id);
