--- src/main/java/com/Teletubbies/Apollo/jwt/test/dto/MemberLoginRequestDto.java
@@ -1,9 +1,0 @@
-package com.Teletubbies.Apollo.jwt.test.dto;
-
-import lombok.Data;
-
-@Data
-public class MemberLoginRequestDto {
-    private String memberId;
-    private String password;
-}
