--- src/main/java/com/Teletubbies/Apollo/jwt/test/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package com.Teletubbies.Apollo.jwt.test.repository;
-
-import com.Teletubbies.Apollo.jwt.test.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByMemberId(String userName);
-}
