--- src/main/java/com/Teletubbies/Apollo/jwt/test/service/CustomUserDetailsService.java
@@ -1,38 +1,0 @@
-package com.Teletubbies.Apollo.jwt.test.service;
-
-import com.Teletubbies.Apollo.jwt.test.domain.Member;
-import com.Teletubbies.Apollo.jwt.test.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailsService implements UserDetailsService {
-    private final MemberRepository memberRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Optional<Member> findMember = memberRepository.findByMemberId(username);
-        System.out.println(findMember.isPresent());
-        return memberRepository.findByMemberId(username)
-                .map(this::createUserDetails)
-                .orElseThrow();
-    }
-    private UserDetails createUserDetails(Member member){
-        return User.builder()
-                .username(member.getUsername())
-                .password(passwordEncoder.encode(member.getPassword()))
-                .roles(member.getRoles().toArray(new String[0]))
-                .build();
-    }
-}
